+++ CHANGELOG.md
@@ -3,6 +3,40 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [1.15.2-1.0.6]
+
+I consider the mod finished. I don't think there will be new features in the future.  
+
+### Added
+- Trapdoor variant for most of the vanilla blocks. To not interfere with vanilla recipes, trapdoor recipes from this mod are like this :  
+```
+###
+   
+###
+```
+Where # is the block you want convert in trapdoor.  
+- Grass Path Slab, Stairs, Fence, Wall and Trapdoor
+- Log and Wood can be stripped to get the stripped version of the block. Work with stairs, slabs, fences, walls and trapdoors.
+- Secret advancements
+- You can convert Coarse Dirt Blocks to Dirt Blocks with a Hoe.
+- Grass Block and Mycelium Block variants can spread to all Dirt Blocks.
+- Iron Ore Stairs & Wall blasting recipe
+- Pumpkin blocks can be sheared to get Carved Pumpkin blocks
+- Ore block variants drop xp
+- Redstone Ore Slabs, Stairs and Fences light up like Redstone Ore
+- Redstone Lamp block variants can now be powered like the vanilla version
+- Netherrack block variants can be smelted to obtain nether brick
+
+### Changed
+- Minimum Forge version : 31.2.0
+- Stairs, slabs, fences and walls variant for leaves block, vine and grass block use overlay/tintindex textures. The color change depending on the biome.
+- Concrete Powder Stairs, Slabs, Walls, Fences and Trapdoors have gravity and turn into Concrete upon touching water
+- Sand and Gravel block variants are affected by gravity
+- Recipes for glass walls don't conflict with glass pane recipe anymore. Recipe for these walls are like the recipes for plank walls
+
+### Removed
+- Ore Slab smelting and blasting recipes
+
 ## [1.15.2-1.0.5]
 ### Added
 - Patchouli guide book. (WIP)